refactor password saving and securing code to a concern: PasswordSecurble
